added flutter svg | updated fulfillment path | minor fulfillment changes
