+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   provider: ^6.0.2
   animated_splash_screen: ^1.1.0
   division: ^0.9.0
+  flutter_svg: ^1.0.1
 
 dev_dependencies:
   flutter_test:
